renamed request spec for import and export of users/teams/topics to better match it's purpose.
